Clean callbacks when unregistering gesture (#3775)
## Description

When using LogicDetector, NativeDetector never cleared the ref to the
methods defined there, firing every version of the callback that was
ever created for a given view. This PR should fix it.

It also removes the redundant `useEffect` with only `viewTag` in its
dependency array - the other one also includes `viewTag` and unregisters
in the `cleanup`.

## Test plan

Uncomment/comment the `onBegin` part of the code and inspect the
`logicMethods` ref in the `NativeDetector`. Before this PR it was never
cleared, after this PR, it should be cleared.

```jsx
  const tapAll = useTap({
    // onBegin: () => {
    //   'worklet';
    // },
    onStart: () => {
      'worklet';
      console.log('Tapped anywhere!');
    },
  });

  const tapFirstPart = useTap({
    onStart: () => {
      'worklet';
      console.log('Tapped on first part!');
    },
  });

  return (
    <View style={styles.container}>
      <NativeDetector gesture={tapAll}>
        <Text style={{ fontSize: 18, textAlign: 'center' }}>
          Some text example running with RNGH
          <LogicDetector gesture={tapFirstPart}>
            <Text style={{ fontSize: 24, color: COLORS.NAVY }}>
              {' '}
              try tapping on this part
            </Text>
          </LogicDetector>
        </Text>
      </NativeDetector>
    </View>
  );
```