Add support for nested Text to LogicDetector (#3778)
## Description

Updates the Gesture Handler backend on the native platforms to support
view hierarchies that get folded into a single host component.

On Android, it relies on
[ReactCompoundView](https://github.com/facebook/react-native/blob/796d182d8989ff826eaa3a57458bdfc79750f820/packages/react-native/ReactAndroid/src/main/java/com/facebook/react/uimanager/ReactCompoundView.kt).
When encountering a view that implements this interface, it's also
hit-tested using its `reactTagForTouch` method, and additional handlers
are used for this particular touch.

On iOS, it relies on
[`touchEventEmitterAtPoint`](https://github.com/facebook/react-native/blob/796d182d8989ff826eaa3a57458bdfc79750f820/packages/react-native/React/Fabric/Mounting/ComponentViews/View/RCTViewComponentView.mm#L1514-L1517)
to achieve a similar effect to the one on Android. The main difference
here is that iOS gesture recognizers need to be attached to a concrete
view in order to be recognized. I made an assumption that any gesture
with `LogicDetector` action type, which is being attached to a
nonexistent view, falls into this category. It will get attached to the
detector component, and a new prop - `virtualHandlerTag` - will be set
to the tag of the "virtual" view. Then, inside
`gestureRecognizerShouldBegin`, `reactTagForTouch` is used to determine
whether the recognizer should be allowed to start processing events or
not.

## Test plan

Added a new screen to the basic example app