[android] fix ripple effect after #3742 (#3781)
## Description

In #3742 I accidentally broke ripple effect, it starts from point of
previous click instead of current (first one starts from the middle).

# Diagnosis and cure

This is because I added a call to `setPressed` in `canBegin`, which
activates the effect before it knows where to start. I added those to
make sure `setPressed` is called somewhere, but after some testing I
think it is not necessary.

## Test plan

<!--
Describe how did you test this change here.
-->