Add automated release scripts (#3784)
## Description

Implements automated release scripts for the repository.

A new nightly release should be triggered on every commit pushed to the
`main` branch. Version resolution happens automatically - the `latest`
tag is pulled from npm, 1 is added to the minor, and the patch is set to
0. The full version is `$major.$minor$.$patch-$year$month$day-$sha`.

To release a stable version, cut a branch with the name matching format
of `$major.$minor-stable` and trigger the `publish-stable-release` on
that branch. By default, it does a dry run, skipping publishing the
package and committing changes. Use that to make sure everything is
correct before an actual release. It also uploads the package that
would've been published as a GitHub artifact, which could be useful for
testing. Version resolution happens automatically:
- the major and minor are inferred from the branch name
- next patch name is figured out from npm, in the dumbest possible way -
iteratively check the patch versions until a not-released one is found
- if the minor is exactly +1 compared to the latest and the patch is 0
or the minor matches the latest and the patch is +1, the release is
tagged as latest

In the future, we may want to run static checks automatically and app
builds before releasing, but that would require refactoring our CI to
rely more on reusable workflows.

## Test plan

Tested on a fork.
Release commitly run:
https://github.com/j-piasecki/react-native-gesture-handler/actions/runs/18902148546/job/53951686147
(failed on publish due to missing provenance)

Release stable run for existing version:
https://github.com/j-piasecki/react-native-gesture-handler/actions/runs/18902287389/job/53952137361
(this one succeeded because I commented out `npm publish` and replaced
it with `echo`)

And here is the created release commit:
https://github.com/j-piasecki/react-native-gesture-handler/commit/211296a9805e68d5e5ae24850c9f03b0a8cfd265.
You can notice that the tag was also created successfully.

Release stable run for a new version:
https://github.com/j-piasecki/react-native-gesture-handler/actions/runs/18903154587/job/53954932559
(this one was a dry run).