Update actions for automatic releases (#3788)
## Description


https://github.com/software-mansion/react-native-gesture-handler/pull/3784
added new release scripts, but there was a leftover workflow for
releasing a commitly. This PR gets rid of it.

I also noticed that the `homepage` field in package.json points to our
readme instead of the docs, so I changed that.

Finally, I added `paths` filter to the release workflow, so changes only
affecting docs, example apps, or scripts won't trigger a release.

## Test plan

The failed job got removed. As for the other change -
https://docs.github.com/en/actions/reference/workflows-and-actions/workflow-syntax#onpushpull_requestpull_request_targetpathspaths-ignore.

> If you define both branches/branches-ignore and paths/paths-ignore,
the workflow will only run when both filters are satisfied.