[Android] Remove workaround for touch events allowing to change state too soon (#3793)
## Description

Supersedes
https://github.com/software-mansion/react-native-gesture-handler/pull/2283

On Android, `onTouchesDown` was delayed to be dispatched after `onBegin`
so that the handlers had the time to initialize themselves (which
happens in `onHandle`) so that the non-touch events would have correct
data.

The old approach was saving the event, which triggered the touch event,
and using that event to initialize the handler during imperative state
change. The issues with that approach were:
1. Storing the event for the duration of the event handling
2. It wouldn't work with asynchronous state changes, which are now
possible

The new approach is to add an option to force gestures to initialize
regardless of the state they are in. This way, when the state is updated
and the current state of the handler is `UNDETERMINED`, the flag is set
on the handler, and when it starts handling the event, it will first
initialize itself.

## Test plan

Tested on the following snippet, with affected gestures
```
import React from 'react';
import { StyleSheet, View } from 'react-native';
import { usePan, NativeDetector } from 'react-native-gesture-handler';

export default function EmptyExample() {
  const pan = usePan({
    onTouchesDown: () => {
      'worklet';
      console.log('Touch down');
      globalThis._setGestureStateSync(57, 4);
    },
    onTouchesMove: () => {
      'worklet';
      console.log('Touch move');
    },
    onTouchesUp: () => {
      'worklet';
      console.log('Touch up');
    },
    onTouchesCancelled: () => {
      'worklet';
      console.log('Touch cancel');
    },
    onBegin: () => {
      'worklet';
      console.log('Gesture begin');
    },
    onStart: (e) => {
      'worklet';
      console.log('Gesture start');
    },
    onUpdate: (e) => {
      'worklet';
      console.log('Gesture update', e.handlerData.translationX, e.handlerData.translationY);
    },
    onEnd: () => {
      'worklet';
      console.log('Gesture end');
    },
    onFinalize: () => {
      'worklet';
      console.log('Gesture finalize');
    },
  });

  console.log('Rendering EmptyExamplee', pan.tag);

  return (
    <View style={styles.container}>
      <NativeDetector gesture={pan}>
        <View style={{ width: 300, height: 300, backgroundColor: 'green' }} />
      </NativeDetector>
    </View>
  );
}

const styles = StyleSheet.create({
  container: {
    flex: 1,
    justifyContent: 'center',
    alignItems: 'center',
  },
});

```