Memoize cloned configs in gesture hooks (#3816)
## Description

Currently config is cloned and modified on every render in gesture
hooks. This PR changes that, so that it's memoized assuming the input is
also stable.

## Test plan

Verify that if the input config is memoized, the returned gesture
reference is stable as long as the config doesn't change (will require
other PRs to be merged first).
-
https://github.com/software-mansion/react-native-gesture-handler/pull/3815
-
https://github.com/software-mansion/react-native-gesture-handler/pull/3812