Hide incorrectly exposed `ReanimatedSwipeable` props. (#3820)
## Description

As stated in [this comment](https://github.com/software-mansion/react-native-gesture-handler/issues/3670#issuecomment-3179955589), some of the `ReanimatedSwipeable` props were incorrectly exposed. Most of them came from `BaseGestureHandlerProps`, but were not actually used in the implementation. 

I've updated `SwipeableProps` type to include only those that were actually used.

I've also moved `hitSlop` from `Animated.View` to `Pan` gesture - it seems that it worked on `iOS`, but not on `android`.

Closes #3670 

## Test plan

Tested on the **"Swipeable reanimation"** example with `hitSlop={-20}` added.