[General] Make `VirtualDetector` part of the public API (#3825)
## Description

As @m-bert pointed out while working on V3 components, the current API
makes certain patterns impossible to accomplish (e.g., `ScrollView` with
`RefreshControl` where `InterceptingGestureDetector` needs to be used)
without forcing the entire subtree to attach gestures to a single
component.

There are two ways to handle that:
- add a mechanism for disabling the existing
`InterceptingGestureDetector` for a given subtree
- stop automatically deciding between `NativeDetector` and
`VirtualDetector` and require users to explicitly use the correct one

The first approach is confusing and not a perfect solution to a problem
of our own making, while we can simply go with the second one, which
gets rid of the problem.

Note: I exported `VirtualDetector` as `VirtualGestureDetector` to stay
consistent.

## Test plan

Updated the example