[General] Allow explicit opt-in for the Animated codepaths when not using native driver (#3824)
## Description

Currently, we're enabling the Animated codepath only when the handler
passed to `onUpdate` is an `Animated.Event` with `useNativeDriver`. When
`useNativeDriver` is not enabled, `Animated.event` returns a function,
and we have no way of differentiating between it and a function passed
by the user.

This is not ideal since the animated codepath receives events in a
different format (wrapped with `nativeEvent`). This meant that the code
would stop working when switching between using the native driver and
not. The same would happen when trying to use a native driver in code
supposed to also run on web, where it's not available - the gesture
would work on mobile, but crash on web.

To avoid this, this PR adds an option to explicitly opt-in to the
animated codepath by setting `useAnimated` in the gesture config. This
will automatically disable Reanimated and throw an error when user if
the user tries to reenable it.

## Test plan

Tested on the `Native Detector` example in the basic app, as well as on
the expo app on the same example for web.