[General] Batch native operations (#3831)
## Description

This takes things a step further than
https://github.com/software-mansion/react-native-gesture-handler/pull/3830.

By wrapping every call to the native module (from V3) with
`NativeProxy`, we have a centralized place that communication with the
module always goes through. Using that, most needed operations are
batched. The exceptions are `updateGestureHandlerConfig` (see comment)
and `createGestureHandler`, which needs to run synchronously. This
should reduce the amount of unnecessary native calls.

On web, "everything is native" so no batching there. We may investigate
the potential gains in the future, but due to differences in scheduling
the same approach as on Android/iOS won't work.

I've also changed `createGestureHandler` on iOS to create the
`UIGestureRecognizer` immediately. This solved a weird "double free"
crash I've been seeing.

## Test plan

Tested on the same code as
https://github.com/software-mansion/react-native-gesture-handler/pull/3830

I saw improvements from ~600ms to ~595ms