[General] Unwrap `handlerData` when passing event to user-defined callbacks (#3840)
## Description

Due to how codegen is structured, we rely on a dynamic `handlerData`
field in events so we don't have to define an event handler for each
event type we may be sending. Currently, this is reflected in the API -
when reading data from events, it needs to be read from `handlerData`.

This PR adds flattened types and a flatten step to get rid of that step,
so that data can be read directly from the event.

This doesn't impact animated, so the event mapping there needs to
include both `nativeEvent` and `handlerData`.

## Test plan

`console.log` the events