[iOS] Fix Animated events with no native driver (#3841)
## Description

I noticed that on iOS, Animated Events with no native driver weren't
working. This was caused by the fact that Animated events were only sent
using the old pipeline, so that NativeAnimatedModule would intercept
them. However, when not using the native driver, the event handler is
just a JS function, which was effectively ignored.

## Test plan

Checked the `NativeDetector` screen with the `useNativeDriver` set to
`false`.