[General] Rename `onStart`/`onEnd` to `onActivate`/`onDeactivate` (#3847)
## Description

It's easy to confuse `onBegin` with `onStart`, `onEnd` with
`onFinalize`, maybe less so, but it can still be confusing. We wanted to
rename `onStart` and the replacement was `onActivate`. To show that
`onEnd` is complementary to that, this PR also changes it to
`onDeactivate`.

Let me know what you think.

## Test plan

Static checks